Move `style.scss` in both webviews to `shared/style.scss`
Also import styles to plots storybook
