+++ webview/src/experiments/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
-import './style.scss'
+import '../shared/style.scss'
 import { App } from './components/App'
 
 const elem = document.createElement('div')
