+++ webview/src/plots/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import './style.scss'
+import '../shared/style.scss'
 import { App } from './components/App'
 
 const elem = document.createElement('div')
