+++ webview/src/shared/style.scss
@@ -15,6 +15,7 @@ body,
   max-width: fit-content;
   width: inherit;
 }
+
 body {
   margin: 0px;
   padding: 0px;
